Django-Middleware-0x03/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
mysql-connector-python==9.3.0
parameterized==0.9.0
propcache==0.3.2
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
twilio==9.6.5
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
